pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Arshan <111618520+ArshansGithub@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
quart = "^0.19.6"
quart-schema = "^0.20.0"
quart-rate-limiter = "^0.10.0"
quart-mongo = "^0.1.2"
pyjwt = "^2.9.0"
quart-cors = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"
pytest = "^8.3.3"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
